Respect external auth sources of portal profiles
When self-registering, we now lookup for the external authentication
source of the current portal profile.

Also, when registering through a sponsor, we set the role and the
unregistration date of the node using the actions of the authentication
source of the current portal profile.. This way, when the sponsor activates
the node, we simply change the node status to "registered".